requirements.txt
wheel
youtube-search-python
simpleeval
wikipedia
aiofiles
vkbottle
qrcode
loguru
psutil
py-cpuinfo
pillow